Add start of result generation tools
It is now called 'db', but name is a bit
confusing maybe.

The goal is to make an easy and clear
module that enables clear overview of
the data that the graphs will be built on.

Each graph should have its data created as a JSON
file so that the data can be verified.